+++ recaf-assembler/src/test/java/me/coley/recaf/assemble/InstructionParseTests.java
@@ -3,7 +3,7 @@ package me.coley.recaf.assemble;
 import me.coley.recaf.assemble.ast.HandleInfo;
 import me.coley.recaf.assemble.ast.insn.*;
 import me.coley.recaf.assemble.parser.BytecodeParser;
-import me.coley.recaf.assemble.parser.BytecodeVisitorImpl;
+import me.coley.recaf.assemble.ast.BytecodeAstGenerator;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -82,7 +82,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnFieldContext fieldCtx = parser.insnField();
 			assertNotNull(fieldCtx, "Parser did not find field context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			FieldInstruction field = (FieldInstruction) visitor.visitInsnField(fieldCtx);
 
 			System.out.println("Original: " + original);
@@ -159,7 +159,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnMethodContext methodCtx = parser.insnMethod();
 			assertNotNull(methodCtx, "Parser did not find method context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			MethodInstruction method = (MethodInstruction) visitor.visitInsnMethod(methodCtx);
 
 			System.out.println("Original: " + original);
@@ -256,7 +256,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnDynamicContext dynamic = parser.insnDynamic();
 			assertNotNull(dynamic, "Parser did not find invoke-dynamic context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			IndyInstruction indy = (IndyInstruction) visitor.visitInsnDynamic(dynamic);
 
 			System.out.println("Original: " + original);
@@ -336,7 +336,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnLdcContext ldcCtx = parser.insnLdc();
 			assertNotNull(ldcCtx, "Parser did not find LDC context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			LdcInstruction ldc = (LdcInstruction) visitor.visitInsnLdc(ldcCtx);
 
 			System.out.println("LDC Value type: " + ldc.getValueType().name());
@@ -370,7 +370,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnIincContext iincCtx = parser.insnIinc();
 			assertNotNull(iincCtx, "Parser did not find IINC context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			IincInstruction insn = (IincInstruction) visitor.visitInsnIinc(iincCtx);
 
 			System.out.println("Original: " + original);
@@ -395,7 +395,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnIntContext intCtx = parser.insnInt();
 			assertNotNull(intCtx, "Parser did not find INT context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			IntInstruction insn = (IntInstruction) visitor.visitInsnInt(intCtx);
 
 			System.out.println("Original: " + original);
@@ -420,7 +420,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnJumpContext jumpCtx = parser.insnJump();
 			assertNotNull(jumpCtx, "Parser did not find JUMP context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			JumpInstruction jump = (JumpInstruction) visitor.visitInsnJump(jumpCtx);
 
 			System.out.println("Original: " + original);
@@ -452,7 +452,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnLineContext lineCtx = parser.insnLine();
 			assertNotNull(lineCtx, "Parser did not find LINE context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			LineInstruction line = (LineInstruction) visitor.visitInsnLine(lineCtx);
 
 			System.out.println("Original: " + original);
@@ -478,7 +478,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnVarContext varCtx = parser.insnVar();
 			assertNotNull(varCtx, "Parser did not find VAR context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			VarInstruction v = (VarInstruction) visitor.visitInsnVar(varCtx);
 
 			System.out.println("Original: " + original);
@@ -503,7 +503,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnTypeContext typeCtx = parser.insnType();
 			assertNotNull(typeCtx, "Parser did not find TYPE context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			TypeInstruction type = (TypeInstruction) visitor.visitInsnType(typeCtx);
 
 			System.out.println("Original: " + original);
@@ -534,7 +534,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnMultiAContext multiCtx = parser.insnMultiA();
 			assertNotNull(multiCtx, "Parser did not find MULTI=ARR context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			MultiArrayInstruction array = (MultiArrayInstruction) visitor.visitInsnMultiA(multiCtx);
 
 			System.out.println("Original: " + original);
@@ -559,7 +559,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnNewArrayContext arrayCtx = parser.insnNewArray();
 			assertNotNull(arrayCtx, "Parser did not find NEWARRAY context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			NewArrayInstruction array = (NewArrayInstruction) visitor.visitInsnNewArray(arrayCtx);
 
 			System.out.println("Original: " + original);
@@ -629,7 +629,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnLookupContext switchCtx = parser.insnLookup();
 			assertNotNull(switchCtx, "Parser did not find LOOKUP-SWITCH context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			LookupSwitchInstruction swtch = (LookupSwitchInstruction) visitor.visitInsnLookup(switchCtx);
 
 			System.out.println("Original: " + original);
@@ -660,7 +660,7 @@ public class InstructionParseTests extends TestUtil {
 			BytecodeParser.InsnTableContext switchCtx = parser.insnTable();
 			assertNotNull(switchCtx, "Parser did not find TABLE-SWITCH context with input: " + original);
 
-			BytecodeVisitorImpl visitor = new BytecodeVisitorImpl();
+			BytecodeAstGenerator visitor = new BytecodeAstGenerator();
 			TableSwitchInstruction swtch = (TableSwitchInstruction) visitor.visitInsnTable(switchCtx);
 
 			System.out.println("Original: " + original);
